- moved usefull functions for : charge, time, amplitude, peak - created simple raw data analysis entry point - moved docopt converts to utils
