+++ digicampipe/calib/camera/charge.py
@@ -1,0 +1,112 @@
+import numpy as np
+from scipy.ndimage.filters import convolve1d
+from digicampipe.utils.utils import get_pulse_shape
+from iminuit import Minuit
+import matplotlib.pyplot as plt
+from probfit import Chi2Regression
+
+
+def compute_charge(events, integral_width, shift, maximum_width=2):
+    """
+
+    :param events: a stream of events
+    :param integral_width: width of the integration window
+    :param maximum_width: width of the region (bin size) to compute charge,
+    maximum value is retained. (not implemented yet)
+    :return:
+    """
+
+    # bins = np.arange(-maximum_width, maximum_width + 1, 1)
+
+    for count, event in enumerate(events):
+
+        adc_samples = event.data.adc_samples
+        pulse_mask = event.data.pulse_mask
+
+        convolved_signal = convolve1d(
+            adc_samples,
+            np.ones(integral_width),
+            axis=-1
+        )
+
+        charges = np.zeros(convolved_signal.shape) * np.nan
+        charges[pulse_mask] = convolved_signal[
+            np.roll(pulse_mask, shift, axis=1)
+        ]
+        event.data.reconstructed_charge = charges
+
+        yield event
+
+
+def compute_amplitude(events):
+
+    for count, event in enumerate(events):
+
+        adc_samples = event.data.adc_samples
+        pulse_indices = event.data.pulse_mask
+
+        charges = np.ones(adc_samples.shape) * np.nan
+        charges[pulse_indices] = adc_samples[pulse_indices]
+        event.data.reconstructed_amplitude = charges
+
+        yield event
+
+
+def fit_template(events, pulse_width=(4, 5), rise_time=12):
+
+    for event in events:
+
+        adc_samples = event.data.adc_samples.copy()
+        time = np.arange(adc_samples.shape[-1]) * 4
+        pulse_indices = event.data.pulse_mask
+        pulse_indices = np.argwhere(pulse_indices)
+        pulse_indices = [tuple(arr) for arr in pulse_indices]
+        amplitudes = np.zeros(adc_samples.shape) * np.nan
+        times = np.zeros(adc_samples.shape) * np.nan
+
+        plt.figure()
+
+        for pulse_index in pulse_indices:
+
+            left = pulse_index[-1] - int(pulse_width[0])
+            left = max(0, left)
+            right = pulse_index[-1] + int(pulse_width[1]) + 1
+            right = min(adc_samples.shape[-1] - 1, right)
+
+            y = adc_samples[pulse_index[0], left:right]
+            t = time[left:right]
+
+            where_baseline = np.arange(adc_samples.shape[-1])
+            where_baseline = (where_baseline < left) + \
+                             (where_baseline >= right)
+            where_baseline = adc_samples[pulse_index[0]][where_baseline]
+
+            baseline_0 = np.mean(where_baseline)
+            baseline_std = np.std(where_baseline)
+            limit_baseline = (baseline_0 - baseline_std,
+                              baseline_0 + baseline_std)
+
+            t_0 = time[pulse_index[-1]] - rise_time
+            limit_t = (t_0 - 2 * 4, t_0 + 2 * 4)
+
+            amplitude_0 = np.max(y)
+            limit_amplitude = (max(np.min(y), 0), amplitude_0 * 1.2)
+
+            chi2 = Chi2Regression(get_pulse_shape, t, y)
+            m = Minuit(chi2, t=t_0,
+                       amplitude=amplitude_0,
+                       limit_t=limit_t,
+                       limit_amplitude=limit_amplitude,
+                       baseline=baseline_0,
+                       limit_baseline=limit_baseline,
+                       print_level=0, pedantic=False)
+            m.migrad()
+
+            adc_samples[pulse_index[0]] -= get_pulse_shape(time, **m.values)
+            amplitudes[pulse_index] = m.values['amplitude']
+            times[pulse_index] = m.values['t']
+
+        event.data.reconstructed_amplitude = amplitudes
+        event.data.reconstructed_time = times
+
+        yield event
