+++ setup.py
@@ -47,6 +47,8 @@ setup(
             'digicam-view=digicampipe.scripts.digicamview:entry',
             'digicam-spe=digicampipe.scripts.spe:entry',
             'digicam-fmpe=digicampipe.scripts.fmpe:entry',
+            'digicam-mpe=digicampipe.scripts.mpe:entry',
+            'digicam-raw=digicampipe.scripts.raw:entry',
             'digicam-rate-scan=digicampipe.scripts.rate_scan:entry',
         ],
     }
